Update README.md
In this study, I investigate change in global causes of death from 1990 to 2019 using data visualization. I begin by highlighting the global disparities in life expectancy, with high life expectancies in regions like Australia, Western Europe, and Canada, while lower life expectancies are common in central and southern regions of Africa. The research aims to assess how causes of death have changed over the years and across different regions.

The dataset used is from the Kaggle platform and contains information on the number of deaths due to various diseases in different countries worldwide.